\name{hwas}
\alias{hwas-package}
\title{Haplotype Wide Association Analysis}
\description{
    Haplotype wide association Analyses (HWAS) are those that find
    statistical association between each haplotype in a population
    and a trait of interest at loci throught throughout the genome.
    This package provides the basic tools for such analyses.
}
\docType{package}
